Move 'select file' stuff to the UI store (#16)
