+++ src/store/reducers/ui.js
@@ -1,5 +1,7 @@
 import { combineReducers } from 'redux';
 
+import { difference } from '../../set';
+
 import { types } from '../actions/ui';
 
 function fileBrowser(state = {}, action) {
@@ -43,6 +45,38 @@ function fileDialog(state = {}, action) {
   }
 }
 
+function selectedFileIds(state = new Set(), action) {
+  switch (action.type) {
+    case types.DESELECT_FILES: {
+      return new Set();
+    }
+    case types.DESELECT_FILES_BULK: {
+      const { ids } = action.payload;
+      return difference(state, ids);
+    }
+    case types.SELECT_FILE: {
+      const { id } = action.payload;
+      return new Set([id]);
+    }
+    case types.SELECT_FILE_ADD: {
+      const { id } = action.payload;
+      const nextState = new Set(state);
+      if (state.has(id)) {
+        nextState.delete(id);
+      } else {
+        nextState.add(id);
+      }
+      return nextState;
+    }
+    case types.SELECT_FILE_BULK: {
+      const { ids } = action.payload;
+      return new Set(ids);
+    }
+    default:
+      return state;
+  }
+}
+
 function scrollOffset(state = {}, action) {
   switch (action.type) {
     case types.SET_SCROLL_OFFSET: {
@@ -74,6 +108,7 @@ function uploader(state = { visibilityFilter: 'all' }, action) {
 export default combineReducers({
   fileBrowser,
   fileDialog,
+  selectedFileIds,
   scrollOffset,
   uploader,
 });
@@ -100,6 +135,11 @@ export const getFileDialogProps = (state, props) => {
   const dialog = state.ui.fileDialog[key];
   return dialog?.props ?? EMPTY_PROPS;
 };
+
+export const getSelectedFileIds = (state) => state.ui.selectedFileIds;
+export const getIsFileSelected = (state, id) => getSelectedFileIds(state).has(id);
+export const getHasSelectedFiles = (state) => getSelectedFileIds(state).size !== 0;
+export const getCountSelectedFiles = (state) => getSelectedFileIds(state).size;
 
 export const getScrollOffset = (state, key) => state.ui.scrollOffset[key] ?? 0;
